Separate functions in positioning code.
